Fixed Shape of LBJ Polygon and added Pleasant Polygon
Spent a lot of time trying to figure out the best way to do overlapping polygons for garages like the Matthew's Street Garage where multiple permits are allowed on different levels. I tried researching stripe fill patterns like are utilized on the Parking Services site, but they use a different overlay API which isn't available for Maps. Solution I came up with is to have an outline color for parking type with the fewest spaces and fill color for the parking type with the most spaces. Also considering adding another smaller polygon in the corner of the larger polygon but that would be a different polygon which would make filtering by color harder.

Also found out that you don't need to add duplicate points for the start and end point of a polygon, it will just autoconnect the last drawn point with the first drawn one.

Used OpenStreetMaps for the points where each lot/garage is located because they have a feature to "square" polygons and Google Maps API does not, so that allows us to have much more even shapes, even if they're a little bit off center from their building outline.

Here is an example of the tool I used:  https://nominatim.openstreetmap.org/search.php?q=Pleasant+Street+Parking+Garage&polygon_geojson=1&viewbox=

Clicking details and then clicking on the way under the OSM row in the details gives a set of nodes that the object is comprised of and clicking on those nodes gives the coordinates.

It's a bit awkward to get these coordinates because they're formatted as links, but by holding alt before clicking them you can copy them without activating the hyperlink. That does download the webpage for me which is a bit weird, but deleting abunch of downloaded pages is a lot faster than fiddling with how you hover the mouse over a hyperlink to select it.

I'll be adding more polygons later today.