Added lots of lots
Added about 20-25 lots covering most of the available parking spots on campus. Also found a tool that makes it a ton easier to get the JS for your polygon. Tool can be found here:
 http://www.birdtheme.org/useful/v3largemap.html

Select polygon and Javascript from the first two dropdown menus to get the JS for a polygon. This saves a TON of time and actually allows you to see what you're drawing while you're drawing it instead of using a mental image. It uses a different format than the rest of our coordinates but that is okay, it still works just the same. It helps to use the satellite view in the top left but that gives you a tilted view. This isn't bad though because you can get more useful views by rotating the map 90 degrees (bottom right) to get a better view. Switching from tilted to top-down gives out-dated low-quality satellite imagery for some reason, so I recommend just using the tilted view and rotating when needed. That can give you a better view of some things anyway.

Also note that this tool won't let you add points convex to the starting point so you have to be careful where you start drawing your polygon. Starting at the innermost corner will ensure that you don't run into this issue in most cases.

See the education lot for an example of the new way to use coordinates.

Also commented out the markers on the map because they were quite large, made the map look cluttered, and probably won't be in the end product anyway.

There still needs to be a few low-capacity lots added but that can be saved for the next sprint.

I will now begin the process of adding data about each lot to a dialog box which becomes visible when the user mouses over the polygon on the map.