Add infowindows to all polygons, optimize polygon creation
Optimize polygon creation to comply with best coding practices by minimizing copy-pasting. Polygons are no longer created one-by-one, but through a for loop. This is achieved by making an array of polygon data for each of the pieces of data that varies between each polygon (coordinates, name, etc.). The data that does not vary per polygon such as stroke weight and opacity is contained in the for loop.

The for loop now also creates an infowindow and event listener for each polygon as well. This infowindow contains the Lot Name, Lot Subtitle, table displaying number of parking spaces based on permit color or parking type (this table will reference the database when it is complete to pull the number of parking spaces), and a blurb describing best times to park at the clicked-on lot.

Edit map dimensions so a larger zoom can be used. The CSS or HTML will have to be changed to recenter the map with its new dimensions.